Upload different environment.yaml file
Environment specific hieradata is meant to be kept in environment.yaml.
In order to not use the development data in production I need to be able
to overwrite it. This commit will allow me ot pass the path of file to
be uploaded over the top of the development environment.yaml